lan96xx: pcie: Remove the SRAM PCIe configuration (UNG_LAGUNA-407)
This lists the configuration for the two SoCs

        LAN966x    LAN969x
---------------------------
Start    1MB        1MB
Size     128KB      2MB

The Synopsis documentation specifies that the BAR start address must be a
multiple of the BAR size, and that is not the case when the start address
is 1MB and the size is 2MB.

So this removes the BAR specifying the SRAM because the address (1MB) used
by the LAN966x SoC is not applicable.

There is also a check for this in the OTP BAR validation function.

Fortunately the LAN969x does not need access to the SRAM as there is to RTE
functionality in this SoC, and that was the main use case for LAN966x
exposing the SRAM via a PCIe BAR.

It is still possible to add this area as 2 1MB BARs by programming the OTP
accordingly.