lan966x: otp: Fix the protection of TBBR
If the ROTPK is not set, then it would enable OTP emulation. So when it
was trying to read again the ROTPK, it would reading a value different
than 0. So it thought that the ROTPK is set, so it would be setting the
OTP_READ_PROTECT.
Fix this by checking if the emulation is enabled or not, as the
emulation is enabled only if ROTPK is different than 0.