lan966x: Update matrix for QSPI0
Currently lan966x allows to access only the first 16MB of the QSPI mmap
area. But in case we have bigger NOR flash on the board then this will
failed with data aboart errors. According to APPL, it says that they
support flashes that have up to 128MB. Therefore update ATF to allow
access to the first 128MB, which correspond also the the maximum allowed
value of matrix.