trng: Fix entropy
It looks like when generating a random number, it is required to pass
number of bits to be generated. The output is stored in uint64_t type.
So when random number is generated, it would zero out all the other bits
that are not used.
For the special case when number of bits is a multiple of uint64_t, it
looks like the last word is zero out. As it would mask the random number
with a mask which will have a value of 0.
The fix here consists in checking if the mask is 0 or not. If it is not
0, then mask the result otherwise don't mask.